mpolset: get/set memory policy and exec a command
mpolset is a standaone executable that can be compiled into a static
binary. In this form, it can be injected into containers, where the
original command of a container can be prefixed with mpolset and
arguments that define the default memory policy for the original
command in the container.

If command is not given, mpolset prints the default memory
policy. This offers an easy way to test how kernels handle missing
nodes and various flag combinations, for instance.

For example: set, execute and get memory policy:

./mpolset -nodes 0-2 -mode MPOL_INTERLEAVE -- ./mpolset

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>